requirements.txt
Pillow>=9.0.0
pyautogui>=0.9.53
pynput>=1.7.6
pywin32>=304
